getPatcoElevatorStatusJson() timestamp
Whoa, interesting. The number of digits in the date and time affected the scrape (e.g., 12:00 vs 1:00). I changed the orientation so that it starts at a fixed point in the string and goes to the end of it, instead of working backwards from the end. Also removed a possible ">" that might show up, and stripped the string so that any leading spaces are removed.